+++ .github/workflows/test.yml
@@ -29,13 +29,13 @@ jobs:
   e2etest:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v7
     - name: Run tests
       run: docker compose -f tests/e2e/docker-compose.yml up --build --exit-code-from pytest
   docs:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v7
       - name: Set up Python 3.11
         uses: actions/setup-python@v5
         with:
