+++ tests/e2e/tests/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+from karton.core.backend import KartonBackend
+from karton.core import Producer, Config
+
+
+@pytest.fixture
+def backend():
+    return KartonBackend(Config())
+
+
+@pytest.fixture
+def producer():
+    return Producer(identity="test-producer")
