+++ tests/e2e/tests/test_errors.py
@@ -3,8 +3,6 @@ import pytest
 
 from shared import (
     BACKENDS,
-    backend,
-    producer,
     wait_for_task_state,
     wait_for_routed_tasks,
 )
