+++ tests/e2e/tests/test_logging.py
@@ -3,8 +3,6 @@ from itertools import islice
 
 from shared import (
     BACKENDS,
-    backend,
-    producer,
     wait_for_routed_tasks,
     wait_for_task_state,
 )
