+++ tests/e2e/tests/test_s3.py
@@ -5,8 +5,6 @@ import os
 
 from shared import (
     BACKENDS,
-    backend,
-    producer,
     wait_for_routed_tasks,
     wait_for_task_state,
 )
